refactor: move shared object loader
